forecast/data-ingestion/requirements.txt
PyGithub
pycoingecko
requests
pandas